PowerTables
Task: Pair program and display a table of powers.

What will the application do?
Prompt the user to enter an integer.
Display a table of squares and cubes from 1 to the value entered.
Ask if the user wants to continue.

Build Specifications
Assume that the user will enter valid data.
Only continue if the user agrees to.